requirements.txt
openai>=1.14.0
tavily-python>=0.5.0
sqlalchemy>=2.0.0
pandas>=2.2.0
streamlit>=1.39.0
plotly>=5.24.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
tenacity>=8.2.0
pydantic>=2.6.0
